Back End/requirements.txt
torch==1.0.1
tqdm==4.43.0
torchvision==0.2.2
dlib==19.6.1
scipy==1.4.1
Flask_SQLAlchemy==2.4.1
Flask==1.1.1
pandas==1.0.1
numpy==1.17.3
opencv_python==3.4.2.16
Scrapy==2.1.0
Flask_Script==2.0.6
selenium==3.141.0
fake_useragent==0.1.11
Werkzeug==1.0.0
youtube_dl==2020.5.3
Flask_Cors==3.0.8
matplotlib==3.2.0
jieba==0.42.1
SQLAlchemy==1.3.13
pretrainedmodels==0.7.4
facenet_pytorch==2.2.9
Pillow==7.1.2
scikit_learn==0.23.1
torchsummary==1.5.1
